% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{generateData}
\alias{generateData}
\title{Generates data-sets from latent factor model}
\usage{
generateData(n, p, k, rep)
}
\arguments{
\item{n}{Numeric, number of data points in each repetition}

\item{p}{Numeric, dimension of each data point}

\item{k}{Numeric, number of factors to generate from}

\item{rep}{Numeric, number of repetitions}
}
\value{
list containing data-sets generated from latent factor model with error variance 0.01, true variance, Lambda and its rank.
}
\description{
Generates data-sets from latent factor model
}
\examples{
newData = generateData( n = 1000, p = 50, k = 10, rep = 5 )
}
