% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSamplerforCov.R
\name{GibbsCov}
\alias{GibbsCov}
\title{Estimates loading matrix and covariance for factor model data through Sparse Bayesian Infinite Factor Model}
\usage{
GibbsCov(data, nrun, burn, thin, epsilon)
}
\arguments{
\item{data}{List, factor model data generated through generateData function}

\item{nrun}{Numeric, number of iterations to run the algorithm}

\item{burn}{Numeric, number of iterations to ignore}

\item{thin}{Numeric, thinning parameter for the outcome}

\item{epsilon}{proper fraction, used in truncation of factors}
}
\value{
List containing covariance matrix, error variance, error is estimating the covariance (if true covariance is supplied), number of chosen factors and posterior estimate of factors accross MCMC iterations.
}
\description{
Estimates loading matrix and covariance for factor model data through Sparse Bayesian Infinite Factor Model
}
\details{
abc ...
}
\examples{
newdata = generateData( n = 100, p = 50, k = 10, rep = 2 )
GibbsCov(data = newdata, nrun = 2000, burn = 500, thin = 3, epsilon = 1e-4)
}
\references{
Bhattacharya et al. (2011). "Sparse Bayesian Infinite Factor Models",
\emph{Biometrika},
\url{https://doi.org/10.1093/biomet/asr013}

Schiavon, Lorenzo, and Antonio Canale. (2020)
"On the truncation criteria in infinite factor models." \emph{Stat}
< https://doi.org/10.1002/sta4.298>
}
