% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datagen.R
\name{imputeData}
\alias{imputeData}
\title{Makes a list from a data table suiatble as an input to GibbsCov. (Imputes the data if necessary)}
\usage{
imputeData(data)
}
\arguments{
\item{data}{data table}
}
\value{
list, containing supplied dataset and other true parameters
}
\description{
Makes a list from a data table suiatble as an input to GibbsCov. (Imputes the data if necessary)
}
\examples{
library(psych)
data = bfi[, 1:25]
imputeData(data)
}
