% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSamplerforCov.R
\name{GibbsCov}
\alias{GibbsCov}
\title{Estimates loading matrix and covariance for factor model data}
\usage{
GibbsCov(data, nrun, burn, thin, epsilon)
}
\arguments{
\item{data}{List, factor model data generated through generateData function}

\item{nrun}{Numeric, number of iterations to run the algorithm}

\item{burn}{Numeric, number of iterations to ignore}

\item{thin}{Numeric, thinning parameter for the outcome}

\item{epsilon}{proper fraction, usd in truncation of factors}
}
\value{
List containing loading matrix, variance and error in estimation
}
\description{
Estimates loading matrix and covariance for factor model data
}
\examples{
# newData = generateData( generateData( n = 1000, p = 50, k = 10, rep = 5 )
Estimates = GibbsCov(data = newData, nrun = 20000, burn = 5000, thin = 3, epsilon = 1e-3)
}
