% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSamplerforCov.R
\name{GibbsCov}
\alias{GibbsCov}
\title{Estimates loading matrix and covariance for factor model data}
\usage{
GibbsCov(data, nrun, burn, thin, epsilon)
}
\arguments{
\item{data}{List, factor model data generated through generateData function}

\item{nrun}{Numeric, number of iterations to run the algorithm}

\item{burn}{Numeric, number of iterations to ignore}

\item{thin}{Numeric, thinning parameter for the outcome}

\item{epsilon}{proper fraction, used in truncation of factors}
}
\value{
List containing loading matrix, variance and error in estimation
}
\description{
Estimates loading matrix and covariance for factor model data
}
\examples{
newdata = generateData( n = 100, p = 50, k = 10, rep = 2 )
GibbsCov(data = newdata, nrun = 2000, burn = 500, thin = 3, epsilon = 1e-4)
}
